[flakey] Deflakey test_ray_shutdown.py (#25422)
The main issue with this test is that the worker is trying to connect to the raylet but the raylet exits, and in this case, it'll hang there. This happens before the periodical check runs so the worker won't exit as well.

This fix moves the hanging part to the place after the periodical check starts.

Another issue is the pubsub timeout. The default one is 60s, and we need to adjust it to smaller value to make it work within 60s for the test.