[Autoscaler] Update default AMIs to latest versions (#25684)
Closes #25588

NVIDIA recently pushed updates to the CUDA image removing support for end of life drivers. Therefore, the default AMIs that we previously had for OSS cluster launcher are not able to run the Ray GPU Docker images.

This PR updates the default AMIs to the latest Deep Learning versions. In general, we should periodically update these AMIs, especially when we add support for new CUDA versions.

I manually confirmed that the nightly Ray docker images work with the new AMI in us-west-2.

Signed-off-by: Shomil Jain <shomil@anyscale.com>