[datasets] Avoid unnecessary metadata serialization in Datasets shuffle (#25734)
Push-based shuffle has some extra metadata involving merge and reduce tasks. Previously we were serializing an O(n) (n = reduce tasks) metadata and sending this to tasks, which caused a lot of unnecessary plasma usage on the head node. This PR splits up the metadata into parts that can be kept on the driver and a relatively cheap part that is sent to all tasks.
Related issue number

One of the issues needed for #24480.

Signed-off-by: Shomil Jain <shomil@anyscale.com>