[Core] install jemalloc in Ray docker and use jemalloc in `benchmark` release tests (#26112)
There are mysterious memory usage growth in Ray clusters that disappear when running with jemalloc. Before we are able to figure out the root cause, it seems using jemalloc by default can be a good walkaround. Because of its efficiency, using jemalloc by default can be beneficial, but we need to run more benchmarks to verify.


Signed-off-by: Shomil Jain <shomil@anyscale.com>